:recycle: public key validation optimisation
